+++ hbw/app/javascript/packs/hbw/components/helpers/index.js
@@ -1,11 +1,13 @@
 import withCallbacks from './callbacks';
 import withDeleteIf from './delete_if';
 import withTasks from './tasks';
+import withSelect from './select';
 import { compose } from './utils';
 
 export {
   withCallbacks,
   withDeleteIf,
   withTasks,
+  withSelect,
   compose
 };