+++ hbw/app/javascript/packs/hbw/components/index.js
@@ -8,7 +8,6 @@ import './error.js';
 import './form.js';
 import './menu.js';
 import './menu_button.js';
-import './mixins.js';
 import './pending.js';
 import './task.js';
 import './task_group.js';